Cool picture
Right before pulling the cylinder head off of a MAN B&W K108ME-C engine to replace a piston, I picked up the giant pipe wrench used to loosen some of the posts on the engine. I used to be an engineer for the US Navy. This is a picture of one of my favorite activities.